Move database processing function into sketchlib.py
